coded the first 2 tasks

+++ Task_2.c
@@ -5,7 +5,7 @@
 The main() function of the following C program initially constructs a binary search tree with each node 
 of type struct node and the data member in each node composed of a positive number greater than zero. 
 After the tree is constructed, the main() function calls a user-defined function delete_tree() that
-takes a pointer to the rootnode of the tree as input argument and deletes all the nodes in the tree. 
+takes the address of a pointer to the rootnode of the tree as input argument and deletes all the nodes in the tree. 
 However, in this case, the main() function utilizes the delete_tree() function to delete all the nodes
 in the right subtree of the rootnode and it does this by calling the function with the rightPtr member
 of the rootnode as input argument.
@@ -31,6 +31,7 @@ the data in the printed output is delimited using one whitespace character.
 */
 
 #include <stdio.h>
+#include <string.h>
 #include <stdlib.h>
 
 struct node
@@ -44,17 +45,27 @@ void insert_node(struct node** treePtr, int data);
 void inOrder(struct node* treePtr);
 void delete_tree(struct node** treePtr);
 
-int main() {
-	int temp = 0;
+int main(int argc, char*argv) 
+{
+	char* tokenPtr = NULL;
+	
+	//int temp = 0;
+	
 	struct node* treePtr = NULL;
-    printf("Enter the value of the new data member: ");
-	scanf("%d", &temp);
-    while (temp > 0)
-    {
-        insert_node(&treePtr, temp);
-        printf("Enter the value of the new data member: ");
-        scanf("%d", &temp);            
-    }
+	tokenPtr = strtok(argv[1],",");
+	while(tokenPtr != NULL)
+	{
+		insert_node(&treePtr,atoi(tokenPtr));
+		tokenPtr = strtok(NULL,",");
+	}
+    //printf("Enter the value of the new data member: ");
+	//scanf("%d", &temp);
+    // while (temp > 0)
+    // {
+    //     insert_node(&treePtr, temp);
+    //     printf("Enter the value of the new data member: ");
+    //     scanf("%d", &temp);            
+    // }
     printf("Initial version of binary tree:\n");
     inOrder(treePtr);
     printf("\n");
@@ -99,7 +110,12 @@ void inOrder(struct node* treePtr)
 
 void delete_tree(struct node** treePtr)
 {
-       free(*treePtr);
-	   delete_tree(&((*treePtr)->leftPtr));
-       delete_tree(&((*treePtr)->rightPtr));
+	if(*treePtr != NULL) // check if the node exists
+	{
+		delete_tree(&((*treePtr)->leftPtr));
+		delete_tree(&((*treePtr)->rightPtr));
+		free(*treePtr); // visit node
+		(*treePtr) = NULL; // inform the tree that node has been freed
+	}
 }
+
